+++ src/data_extraction.py
@@ -1,7 +1,5 @@
 from typing import List, Dict
-from openai import OpenAI
-from dotenv import load_dotenv
-from iem_model import AbstractAppConfig, AppModel
+from app_model import AppModel
 import json
 from llm_service import LLM, GPT4o
 from error_handling import ValidationException
