+++ src/llm_service.py
@@ -1,16 +1,12 @@
 from dotenv import load_dotenv
 from pydantic.dataclasses import dataclass
-import instructor
 from openai import OpenAI
 from openai.types.chat import ChatCompletion, ChatCompletionMessageToolCall
 import os
-import streamlit as st
-from enum import Enum
-import json
 from error_handling import LLMInteractionException
 
-from iem_model import AppModel
-from abc import ABC, abstractmethod
+from app_model import AppModel
+from abc import ABC
 from typing import List, Dict, Optional
 
 load_dotenv()
