+++ src/streamlit_app.py
@@ -27,15 +27,17 @@ for message in st.session_state.messages:
 if prompt := st.chat_input("Write something"):
     # Calling the LLM and possibly change values
     with st.chat_message("user"):
-     st.markdown(prompt)
-    
+        st.markdown(prompt)
+
     st.session_state.messages.append({"role": "user", "content": prompt})
 
-    response_message, current_model = strategy.send_message(prompt, st.session_state.messages)
+    response_message, current_model = strategy.send_message(
+        prompt, st.session_state.messages
+    )
 
     with st.chat_message("assistant"):
         st.markdown(response_message)
-        
+
     st.session_state.messages.append({"role": "assistant", "content": response_message})
 
     # TODO: Add an potential extra system promt to st.session_state.messages to tell the LLM
@@ -44,8 +46,6 @@ if prompt := st.chat_input("Write something"):
     with st.sidebar:
         st.subheader("Configuration Parameters")
         st.markdown(current_model.generate_prompt_sidebar())
-
-
 
 
 st.divider()
